Split QualifiedNameProvider out of scope_provider.py
