+++ libcst/metadata/__init__.py
@@ -16,6 +16,9 @@ from libcst.metadata.expression_context_provider import (
     ExpressionContextProvider,
 )
 from libcst.metadata.full_repo_manager import FullRepoManager
+from libcst.metadata.name_provider import (
+    QualifiedNameProvider,
+)
 from libcst.metadata.parent_node_provider import ParentNodeProvider
 from libcst.metadata.position_provider import (
     PositionProvider,
@@ -37,7 +40,6 @@ from libcst.metadata.scope_provider import (
     FunctionScope,
     GlobalScope,
     QualifiedName,
-    QualifiedNameProvider,
     QualifiedNameSource,
     Scope,
     ScopeProvider,
